refactor(web): migrate cms preview netlify function to typescript
