+++ package.json
@@ -61,6 +61,7 @@
     "prettier": "^3.2.5"
   },
   "devDependencies": {
+    "@netlify/functions": "^2.6.0",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@typescript-eslint/eslint-plugin": "^7.1.1",
     "@typescript-eslint/parser": "^7.1.1",
